feat <김상현 #9> update wrong directory name
