+++ src/main/java/com/sc/surveymarket/SurveyMarketApplication.java
@@ -1,9 +1,8 @@
-package com.sc.suverymarket;
+package com.sc.surveymarket;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -12,10 +11,10 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableMongoRepositories
 @EnableScheduling
 @SpringBootApplication(exclude = DataSourceAutoConfiguration.class)
-public class SuveryMarketApplication {
+public class SurveyMarketApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SuveryMarketApplication.class, args);
+		SpringApplication.run(SurveyMarketApplication.class, args);
 	}
 
 }
