+++ src/main/java/com/sc/surveymarket/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.sc.suverymarket.config;
+package com.sc.surveymarket.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
