+++ src/main/java/com/sc/surveymarket/document/Survey.java
@@ -1,45 +1,43 @@
-package com.sc.suverymarket.document;
+package com.sc.surveymarket.document;
 
+import com.sc.surveymarket.document.question.Question;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Document
 @Getter
 @NoArgsConstructor
-public class User {
+public class Survey {
 
     @Id
-    @Column(name = "user_id")
+    @Column(name = "survey_id")
     @GeneratedValue(strategy = IDENTITY)
     private String id;
 
-    @Column(nullable = false, length = 20)
-    private String nickname;
+    @Column(nullable = false)
+    private User user;
 
-    @Column(nullable = false, length = 83)
-    private String email;
-
-    @Column(nullable = false, length = 13)
-    private String phoneNumber;
+    @Column(nullable = false, length = 100)
+    private String title;
 
     @Column(nullable = false)
-    private String socialId;
+    private List<Question> questionList;
 
-    @CreatedDate
+    @CreatedDate // Insert
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
 
-    @LastModifiedDate
+    @LastModifiedDate // Insert or Update
     @Column(nullable = false)
     private LocalDateTime modifiedAt;
 
@@ -47,11 +45,9 @@ public class User {
     private Integer version;
 
     @Builder
-    public User(String id, String nickname, String email, String phoneNumber, String socialId) {
-        this.id = id;
-        this.nickname = nickname;
-        this.email = email;
-        this.phoneNumber = phoneNumber;
-        this.socialId = socialId;
+    public Survey(User user, String title, List<Question> questionList) {
+        this.user = user;
+        this.title = title;
+        this.questionList = questionList;
     }
 }
