+++ src/main/java/com/sc/surveymarket/dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package com.sc.suverymarket.dto;
+package com.sc.surveymarket.dto;
 
 import lombok.*;
