+++ src/main/java/com/sc/surveymarket/repository/SurveyRepository.java
@@ -1,6 +1,6 @@
-package com.sc.suverymarket.repository;
+package com.sc.surveymarket.repository;
 
-import com.sc.suverymarket.document.Survey;
+import com.sc.surveymarket.document.Survey;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface SurveyRepository extends MongoRepository<Survey, String> {
