+++ src/main/java/com/sc/surveymarket/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.sc.suverymarket.repository;
+package com.sc.surveymarket.repository;
 
-import com.sc.suverymarket.document.User;
+import com.sc.surveymarket.document.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface UserRepository extends MongoRepository<User, String> {
