+++ src/test/java/com/sc/surveymarket/SuveryMarketApplicationTests.java
@@ -1,6 +1,5 @@
-package com.sc.suverymarket;
+package com.sc.surveymarket;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
